feature: Separando uma branch específica p/ cypress com cenários em formato ação/resultado esperado
